backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
aiogram==3.13.1
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.6.1
alembic==1.13.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.1
python-multipart==0.0.9
redis==5.0.1
celery==5.3.4
pytz==2024.1
aiohttp==3.9.5